moving poll ui to it's own file
